pyproject.toml
[tool.poetry]
name = "tap-keyingress_spss"
version = "0.0.1"
description = "`tap-keyingress_spss` is a Singer tap for keyingress_spss, built with the Meltano SDK for Singer Taps."
authors = ["Dirk Heinecke"]
keywords = [
    "ELT",
    "keyingress_spss",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = ">=3.7.1,<3.11"
requests = "^2.25.1"
singer-sdk = "^0.3.17"
pandas = "^1.2.0"
pyreadstat = "^1.1.4"
google-auth = "^2.3.3"
google-cloud-storage = "^1.43.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "^21.9b0"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "tap_keyingress_spss"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-keyingress_spss = 'tap_keyingress_spss.tap:Tapkeyingress_spss.cli'
